# Optimize Char#to_s