--- a/src/char.cr
+++ b/src/char.cr
@@ -672,7 +672,7 @@ struct Char
   # 130
   # ```
   def each_byte : Nil
-    # See http://en.wikipedia.org/wiki/UTF-8#Sample_code
+    # See https://en.wikipedia.org/wiki/UTF-8#Description
 
     c = ord
     if c < 0x80
@@ -705,7 +705,7 @@ struct Char
   # '好'.bytesize # => 3
   # ```
   def bytesize
-    # See http://en.wikipedia.org/wiki/UTF-8#Sample_code
+    # See https://en.wikipedia.org/wiki/UTF-8#Description
 
     c = ord
     if c < 0x80
@@ -746,7 +746,7 @@ struct Char
   # 'あ'.to_s # => "あ"
   # ```
   def to_s
-    String.new(4) do |buffer|
+    String.new(bytesize) do |buffer|
       appender = buffer.appender
       each_byte { |byte| appender << byte }
       {appender.size, 1}