--- a/src/char.cr
+++ b/src/char.cr
@@ -752,10 +752,15 @@ struct Char
   # 'あ'.to_s # => "あ"
   # ```
   def to_s : String
-    String.new(bytesize) do |buffer|
-      appender = buffer.appender
-      each_byte { |byte| appender << byte }
-      {appender.size, 1}
+    chars = uninitialized UInt8[4]
+    i = 0
+    each_byte do |byte|
+      chars[i] = byte
+      i += 1
+    end
+    String.new(i) do |buffer|
+      buffer.copy_from(chars.to_unsafe, i)
+      {i, 1}
     end
   end