This optimizes [`Char#to_s`](https://crystal-lang.org/api/0.27.2/Char.html#to_s-instance-method) by passing the exact amount of bytes to [`String.new`](https://crystal-lang.org/api/0.27.2/String.html#new%28capacity%3AInt%2C%26block%29-class-method) so if a character has less than 4 bytes (which is very commonly the case), less memory will be allocated by [`GC.malloc_atomic`](https://crystal-lang.org/api/0.27.2/GC.html#malloc_atomic(size:Int)-class-method) which is faster.
1 byte character:
```
new to_s  64.32M ( 15.55ns) (±14.73%)  16 B/op        fastest
old to_s  23.16M ( 43.18ns) (±21.06%)  48 B/op   2.78× slower
```
2 byte character:
```
new to_s  47.31M ( 21.14ns) (±20.04%)  16 B/op        fastest
old to_s  22.29M ( 44.86ns) (±22.60%)  48 B/op   2.12× slower
```
3 byte character:
```
new to_s  47.76M ( 20.94ns) (±10.11%)  32 B/op        fastest
old to_s  31.36M ( 31.89ns) (± 8.03%)  32 B/op   1.52× slower
```
4 byte character, the new `to_s` is slightly slower than the old `to_s`:
```
new to_s  45.33M ( 22.06ns) (±12.35%)  32 B/op   1.10× slower
old to_s  49.89M ( 20.05ns) (± 7.99%)  32 B/op        fastest
```
I believe that is because in the new `to_s` the `4` needs to be computed first using [`Char#bytesize`](https://crystal-lang.org/api/0.27.2/Char.html#bytesize-instance-method) while the old `to_s` has the `4` already.
But I think it's worth it.

This also fixes two dead links.