@asterite 2019-02-07
@r00ster91 Thank you! This is a good idea. However, calling `bytesize` and then `each_byte` effectively decodes the char twice, that's why with 1 byte it's a bit slower.

Using your initial idea of avoiding allocating more memory than needed, I ended up with this implementation:

```crystal
def new_to_s
  chars = uninitialized UInt8[4]
  i = 0
  each_byte do |byte|
    chars[i] = byte
    i += 1
  end
  String.new(i) do |buffer|
    buffer.copy_from(chars.to_unsafe, i)
    {i, 1}
  end
end
```

Using this benchmark:

```crystal
require "benchmark"

Benchmark.ips do |x|
  x.report("1 (old)") do
    'a'.to_s
  end
  x.report("1 (new)") do
    'a'.new_to_s
  end
end

Benchmark.ips do |x|
  x.report("2 (old)") do
    'ߣ'.to_s
  end
  x.report("2 (new)") do
    'ߣ'.new_to_s
  end
end

Benchmark.ips do |x|
  x.report("3 (old)") do
    '勞'.to_s
  end
  x.report("3 (new)") do
    '勞'.new_to_s
  end
end

Benchmark.ips do |x|
  x.report("4 (old)") do
    '䈽'.to_s
  end
  x.report("4 (new)") do
    '䈽'.new_to_s
  end
end
```

I got these results:

```
1 (old)  19.64M ( 50.91ns) (± 1.84%)  48 B/op   1.89× slower
1 (new)  37.07M ( 26.98ns) (± 1.72%)  16 B/op        fastest
2 (old)  19.76M (  50.6ns) (± 2.96%)  48 B/op   1.99× slower
2 (new)  39.26M ( 25.47ns) (± 3.33%)  16 B/op        fastest
3 (old)  17.45M ( 57.31ns) (± 3.39%)  33 B/op   1.46× slower
3 (new)  25.56M ( 39.12ns) (± 2.62%)  32 B/op        fastest
4 (old)  17.27M ( 57.89ns) (± 4.55%)  33 B/op   1.45× slower
4 (new)   25.0M ( 39.99ns) (± 6.45%)  32 B/op        fastest
```

(always faster than before)

Can you run this benchmark in your machine to check it's also always faster?