@r00ster91 2019-02-07
Oh so in your implementation you avoid some of these checks: `elsif c <= 0xffff` by not calling both `bytesize` and `each_byte` like in the implementation of this PR, that's a good idea.

> that's why with 1 byte it's a bit slower.

I think you mean 4 bytes, right?

In your last benchmark the `䈽` isn't 4 bytes, it's 3 bytes. When I replace it with `􏾛` which has 4 bytes, these are my results:
```
1 (old)  20.95M ( 47.73ns) (±21.06%)  48 B/op   3.33× slower
1 (new)  69.83M ( 14.32ns) (±13.88%)  16 B/op        fastest
2 (old)  19.95M ( 50.13ns) (±21.08%)  48 B/op   2.88× slower
2 (new)  57.51M ( 17.39ns) (±19.27%)  16 B/op        fastest
3 (old)  24.89M ( 40.18ns) (±14.46%)  32 B/op   2.09× slower
3 (new)  52.07M ( 19.21ns) (± 8.31%)  32 B/op        fastest
4 (old)  50.25M (  19.9ns) (± 8.13%)  32 B/op        fastest
4 (new)  50.12M ( 19.95ns) (± 7.47%)  32 B/op   1.00× slower
```
So in the last benchmark the new implementation is at least not slower.

But now when I also benchmark with the `to_s` of this PR:
```crystal
require "benchmark"

struct Char
  def new_to_s
    chars = uninitialized UInt8[4]
    i = 0
    each_byte do |byte|
      chars[i] = byte
      i += 1
    end
    String.new(i) do |buffer|
      buffer.copy_from(chars.to_unsafe, i)
      {i, 1}
    end
  end
  def pr_to_s
    String.new(bytesize) do |buffer|
      appender = buffer.appender
      each_byte { |byte| appender << byte }
      {appender.size, 1}
    end
  end
end

Benchmark.ips do |x|
  x.report("1 (old)") do
    'a'.to_s
  end
  x.report("1 (new)") do
    'a'.new_to_s
  end
  x.report("1 (pr)") do
    'a'.pr_to_s
  end
end
Benchmark.ips do |x|
  x.report("2 (old)") do
    'ߣ'.to_s
  end
  x.report("2 (new)") do
    'ߣ'.new_to_s
  end
  x.report("2 (pr)") do
    'ߣ'.pr_to_s
  end
end
Benchmark.ips do |x|
  x.report("3 (old)") do
    '勞'.to_s
  end
  x.report("3 (new)") do
    '勞'.new_to_s
  end
  x.report("3 (pr)") do
    '勞'.pr_to_s
  end
end
Benchmark.ips do |x|
  x.report("4 (old)") do
    '􏾛'.to_s
  end
  x.report("4 (new)") do
    '􏾛'.new_to_s
  end
  x.report("4 (pr)") do
    '􏾛'.pr_to_s
  end
end
```
I get this where the PR implementation is a tiny bit faster than your implementation for some reason:
```
1 (old)  17.03M ( 58.72ns) (±20.07%)  48 B/op   4.08× slower
1 (new)  68.56M ( 14.58ns) (±12.98%)  16 B/op   1.01× slower
 1 (pr)  69.44M (  14.4ns) (±11.49%)  16 B/op        fastest
2 (old)  22.14M ( 45.16ns) (±22.05%)  48 B/op   3.07× slower
2 (new)  65.38M (  15.3ns) (±12.42%)  16 B/op   1.04× slower
 2 (pr)  67.88M ( 14.73ns) (±13.54%)  16 B/op        fastest
3 (old)  29.58M ( 33.81ns) (± 8.87%)  32 B/op   1.80× slower
3 (new)  52.56M ( 19.02ns) (± 7.47%)  32 B/op   1.01× slower
 3 (pr)  53.34M ( 18.75ns) (± 7.06%)  32 B/op        fastest
4 (old)  28.27M ( 35.37ns) (±12.94%)  32 B/op   1.87× slower
4 (new)  52.58M ( 19.02ns) (± 7.87%)  32 B/op   1.01× slower
 4 (pr)  52.93M ( 18.89ns) (± 7.48%)  32 B/op        fastest
```
Hmm but I'm not sure if there is actually really a difference.
Now only comparing those 3:
```crystal
Benchmark.ips do |x|
  x.report("4 (old)") do
    Char::MAX.to_s
  end
  x.report("4 (new)") do
    Char::MAX.new_to_s
  end
  x.report("4 (pr)") do
    Char::MAX.pr_to_s
  end
end
```
```
4 (old)  45.47M ( 21.99ns) (±12.17%)  32 B/op   1.10× slower
4 (new)   49.9M ( 20.04ns) (± 8.18%)  32 B/op        fastest
 4 (pr)  49.23M ( 20.31ns) (± 8.60%)  32 B/op   1.01× slower
```
Yes it looks like they are relatively the same.