@r00ster91 2019-02-07
In the 4 byte character benchmark in the description of this PR it was just a coincidence I think. I was just sure the new implementation was slower in this case because I've ran it multiple times in a row and it was always shown as 'slower'. But now I also got a case where the old one was 'slower':
```shell
$ crystal benchmark.cr --release
4 (pr)   44.4M ( 22.52ns) (±14.21%)  32 B/op   1.13× slower
4 (old)  50.39M ( 19.85ns) (± 8.81%)  32 B/op        fastest
$ crystal benchmark.cr --release
4 (pr)   23.81M ( 41.99ns) (±14.00%)  32 B/op   1.94× slower
4 (old)  46.09M (  21.7ns) (±20.27%)  32 B/op        fastest
$ crystal benchmark.cr --release
4 (pr)   47.82M ( 20.91ns) (±12.49%)  32 B/op        fastest
4 (old)  41.89M ( 23.87ns) (±23.95%)  32 B/op   1.14× slower
```