@r00ster91 2019-03-17
I resolved the conflicts.