@straight-shoota 2019-03-17
I think the implementation should be changed to @asterite's proposal in https://github.com/crystal-lang/crystal/pull/7393#issuecomment-461575678

It seems a bit weird that in your benchmark, it was en par with the `bytesize` implementation, but in theory it should be more performant.