@r00ster91 2021-04-27
@straight-shoota Here's another benchmark.
```cr
struct Char
  def new_to_s
    chars = uninitialized UInt8[4]
    i = 0
    each_byte do |byte|
      chars[i] = byte
      i += 1
    end
    String.new(i) do |buffer|
      buffer.copy_from(chars.to_unsafe, i)
      {i, 1}
    end
  end
end

require "benchmark"

Benchmark.ips do |benchmark|
  benchmark.report("to_s") { 'あ'.to_s(); 'a'.to_s(); }
  benchmark.report("new_to_s") { 'あ'.new_to_s(); 'a'.new_to_s(); }
end
```
```
    to_s  14.87M ( 67.25ns) (±16.56%)  80.0B/op   2.91× slower
new_to_s  43.29M ( 23.10ns) (±22.26%)  48.0B/op        fastest
```
`new_to_s` is still consistently faster for me and allocates less memory. Same for the benchmark @asterite provided above.
Is anything blocking this?